requirements.txt
streamlit
pyrebase4
python-dotenv
setuptools
streamlit-option-menu
requests
